WIP commit, append only, (level to compact in the end)
With level, deletes are writes. So why not do everything at the vnode
level as a write, and use level to compact away tomsbtones and
dominated keys?

Each set is stored in level like this

{s, SetName :: binary(), clock} -> Clock :: bigset_clock:clock().
A first key that stores a clock, this clock is a VV with exceptions
right now, and needs work.

Each element is stored like this

{s, SetName::binary(), Element::binary(), Actor::binary(), Counter::pos_integer(), TSB=0|1}

What this means is that every ADD of an element adds a key that is made
up of the set name, element, and dot at which it was added. The
last element in the key is a Tombstone Bit. 0 for added, 1 for removed,
just like the old orset.

This means when some vnode is told to remove <<"X">> with Ctx [{a, 4}, {b, 6}]
rather than reading <<"X">> and checking it's dot, we just write two
keys

{s, Set, <<"X">>, a, 4, 1}
{s, Set, <<"X">>, b, 6, 1}

If <<"X">> exists at the vnode at a dot {b, N} where N =< 6 then
this new key is a tomstone. It is safe, even if b never wrote <<"X">> at 6,
since any <<"X">> written by b is either =< 6 or > 6. The latter dominates, the
former is dominated by the tombstone. That extrinsic event set shit from
the WinFS CVE paper!

This scheme needs a clever fold in the vnode when reading. And
a streaming merge at the coordinator when reading. It also needs a read
of the clock before writing. But that is the only read, so it should
compare very well with riak_kv sets, especially for larger sets.

NOTE: the future plan is to have level compact out the extra info
and tombstones. Example

{s, S, clock} -> [{a, 10}, {b, 3}]
{s, S, X, a, 4, 0}
{s, S, X, a, 6, 0}
{s, S, X, a, 7, 1}
{s, S, X, b, 1, 0}
{s, S, X, b, 2, 0}

Would compact to

{s, S, clock} -> [{a, 10}, {b, 3}]
{s, S, X, b, 2, 0}

By this rule: for any key {s, S, X, A, N, 0} discard if somekey
{s, S, X, A, N', 0} exists where N' > N.
For any key {s, S, X, A, N, 0} discard if some key
{s, S, X, A, N, 1} exists
For any key
{s, S, X, A, N, 1} discard if contiguous portion of clock (the VV)
descends {A, N}. That way we can drop dominated writes and tombstones.