WIP fix the read errors
Turns out that tuple size is included in sorts!! So the key scheme of
{s, Set, clock} followed by
{s, Set, Element, Actor, Cnt, TSB} meant all the clocks were clumped
together. Oops. Main change was to make the clock key as long as
the element key.

Got a nice basic benchmark on local machine going. Needs R installed
and ggplot and getopt. And 4 nodes of riak running. And 4 nodes of bigset.