Start work for query commans `contains`
If sets are big it makes sense that rather than reading the whole set
instead we might want to query it. The simplest query would be `contains
X`. With bigset decomposed design this should be a lot quicker as it
should read only two keys, the clock and the element. And do a partial
merge on them.