WIP: Add a contains query
Why bring the whole set to the client just find out if 'X' is a member?

For a single element this should be a very simple thing, for multiple
elements it get's difficult: when is it cheaper to seek, seek, seek vs
fold over the key space between min(query_members) and
max(query_members), only accumulating the values we need?