On remove, add elements dots to the clock
When removing some element(s), each element comes with a context of
dots, previously we used these dots to create tombstone writes, but did
not add the dots to the Set clock. Since receiving these removes is the
same as seeing the adds AND removes, we need to add the dots to the
clock.