Add explicit end_key for streaming folds
Streaming folds are more efficient with and end_key as they read from a
start key upto some large buffer OR on end key. Adding an explicit end
key meant fiddling with the key scheme some more to add a single byte
key type, I took the chance to make the tombstone a single byte too.

$c means a clock key
$e means an element key
$z means an end key

Inside the key $a for the TSB is an add
$r for the TSB is a remove