Update EQC for action at a distance
The previous iteration of the code had the same action-at-a-distance bug
as riak data types for the following simple trace

Client A adds X and it goes to node M
Client B adds X and it also goes to node M

The second add effectively _removes_ the first, even though it is
technically concurrent. Imagine

M replicates X {a, 1} to N
M replicated X {a, 2} to O

A read of C from O and subsequent remove would have the effect of
removing X {a, 1} even though it was not seen. Action At A Distance
means you can't use the context at vnodes for adds, but the context as
seen by clients.

TODO: add the code the bigsets itself.
Come up with optional defaults for clients that don't want to read(X)
before adding X (default to [] for safety or vnode_ctx for danger)