Add a context to each write
Action at a distance requires we track extra causal information for each
write. As the previous commit message describes, two clients adding 'X'
concurrently at node B should result in 'X' in the set in such a way
that removing 'X' {B, 2} will not remove 'X' @ {B, 1} unless it is also
seen, in other words, we need a way to treat dots {B, 1} and {B, 2} as
concurrent. The DVV is the way to do that. Store the add/remove Ctx as
the insert/remove payload against the dot of the insert/remove. Thus
`{'X', B, 1, ?ADD}` -> [] is concurrent with `{'X', B, 2, ?ADD} -> []
and a subsequent remove `{'X', A, 1, ?REMOVE} -> {[], [{B, [2]}]}` will
only remove `{'X', B, 2, ?ADD}`.