Handoff should not be full state read at receiver
When a vnode receives handoff it should not have to read all the local
state. The commit adds a handoff technique that means only reading the
clock at the receiving end. Although for the eqc one last part is
missing and we do filter the receiving keys (next commit to address).

The scheme works like this: when handoff starts the receiving node reads
it's clock, and creates a `tracking clock` that is a fresh clock. Each
key it receives is added to the `tracking clock`. When all keys are sent
from the handing off node the handing off node's clock is compared to
the tracking clock. Any events in the From clock but not in the tracking
clock are keys the sender removed (i.e. the trackers complement in the
sender's set of events), we call these the deleted set of events. (NOTE:
the tracking clock is always a subset of the sender clock!) Any events
that are in the receiving clock BEFORE it started getting handoff, but
are not in the deleted set of events (i.e. the intersection of receiving
clock and deleted set of events) need to be removed from the
receiver. How will be addressed in a later commit.

Finally merge the sender clock with the receiver clock to ensure keys
removed by the sender and unseen by the receiver never get added to the
receiver.