Don't "read" handoff receiver keys, write a filter
"Filter" is a terrible name, but rather than reading keys at the
receiver, write a clock that is a read filter. Then at compaction time
remove keys covered by that filter, and at read ignore keys
covered. These keys are the ones that handoff tells us are removed

TODO trim the filter as you remove dots it covers (aim to keep it
small/non-existing)
TODO this is currently failing, counter example is committed