Keep the handoff filter thingy small as possible
When the handoff filter does it's job and causes compaction to remove a
key, the handoff filter can "forget" about that dot. This commit enables
removing a dot from the handoff filter. Since the handoff filter is a
clock, this basically means allowing a clock to "un-see" some event.

NOTE: depends on an API addition in riak_dt_vclock:set_counter/3