Add Handoff code to PoC
Handoff of bigsets is different to regular riak. In riak each key/value
handed off is a complete datum, with bigsets each key/value is part of
some larger whole.

This commit adds code as tested in the bigset_eqc/model code. Handoff
from the sender is basically the same. The difference is that the
receiver needs to know _who_ the sender is, to that end the sender
prepends their `vnodeid` to each message.

The receiver needs to `merge` the incoming bigset with their own, but we
would like to avoid reading the whole set from the receiving vnode. To
that end the receiver keeps some state in memory for each vnode that is
handing off to it (yes, many vnodes may handoff to the receiver at the
same time, but the vnode's serial operation means these are interleaved,
not concurrent operations.)

Since the sender folds in order over it's contents and sends them, we
can assume that the receiver receives an ordered set of messages,
starting with keys for set A, then elements, and finally an end key,
followed by the same for set B etc. If handoff encounters an error, it
starts from the beginning.

Per set received the receiver does the following: when it receives a
clock key it sets the inmemory `tracker` to a fresh clock. When it
receives the sender's clock key it adds it to a local state
`sender_clock` When it receives an element key, it reads the dot for the
key into the `tracker`. When it receives an end key it generates a
filter from the tracker and stores it, and clears in memory state.

The filter is clock that contains the set of events that is the
intersection of events the sender has removed, and the receiver has
seen. The filter is used to ignore keys on fold (read/handoff), and
ultimately to remove keys at compaction.