WIP: move to a design that requires a per-elem-ctx
To simplify the design, read, and lower storage usage, require that
every operation (Add | Remove) has a context. The context is
per-element. On read we return an orswot. A list of pairs
{Element::binary(), Context::binary()} (as of this commit that Context
is just list of dots, not a binary.) If the client wishes to add element
X they first check with `is_member(X)` (not return yet!) And if the
answer is {true, Ctx()} and they still wish to add the member, they must
say {add, {X, Ctx()}} to ensure adds are removes
deterministically. Likewise for removes.