remove the conflated causal
Scott rightly pointed out that having the tombstone for removes is
a bytes in/out quadratic thing. And maybe if we have the information
just issue leveldb deletes. This branch is for that, this commit
is WIP step one (it is a breaking commit, code will not compile)
while I switch branches.