Add compile time key modules
In order to compare A to B key schemes add a behaviour for keys modules
that itself delegates to a macro declared concrete key providing
module. Also add two schemes: lens, which uses typical erlang encoding
but depends on the bigset key comparator, and nulls, which can be used
with the default leveldb key comparator. NOTE: if you write with keys
from one module, you must read with that module. This is NOT production
level inter-op code, just an easy way to make a pair of branches to test against.